MDEV-11588 Support for ONLY_FULL_GROUP_BY functional dependency
Implement a simple case when INNER join is used.
Support cases when primary key or unique keys are used in GROUP BY (then
all fields of this table can be used in SELECT list).
Support a case when virtual column is used. If it is defined with GROUP BY
fields then it can be used in SELECT list.